added select menu for genres, made button to go to addmovie page, changed the name of some files
